fix: move pokemonSelectionSlice to store folder and refactor
