+++ src/components/store/pokemonSelectionSlice.ts
@@ -1,12 +1,12 @@
 import { createSlice, type PayloadAction } from '@reduxjs/toolkit';
-import type { Pokemon } from '../components/PokemonSearch/types/pokemonSearch.types';
+import type { Pokemon } from '../PokemonSearch/types/pokemonSearch.types';
 
 export interface PokemonSelectionState {
-  selectedPokemons: Record<number, Pokemon>;
+  selectedPokemons: Pokemon[];
 }
 
 const initialState: PokemonSelectionState = {
-  selectedPokemons: {},
+  selectedPokemons: [],
 };
 
 export const pokemonSelectionSlice = createSlice({
@@ -16,21 +16,16 @@ export const pokemonSelectionSlice = createSlice({
     togglePokemonSelection: (state, action: PayloadAction<Pokemon>) => {
       const pokemon = action.payload;
 
-      if (state.selectedPokemons[pokemon.id]) {
-        state.selectedPokemons = Object.fromEntries(
-          Object.entries(state.selectedPokemons).filter(
-            ([id]) => id !== pokemon.id.toString()
-          )
+      if (state.selectedPokemons.find((item) => item.id === pokemon.id)) {
+        state.selectedPokemons = state.selectedPokemons.filter(
+          (item) => item.id !== pokemon.id
         );
       } else {
-        state.selectedPokemons = {
-          ...state.selectedPokemons,
-          [pokemon.id]: pokemon,
-        };
+        state.selectedPokemons = [...state.selectedPokemons, pokemon];
       }
     },
     clearAllSelections: (state) => {
-      state.selectedPokemons = {};
+      state.selectedPokemons = [];
     },
   },
 });
